Get flashcards to hilight when hovering over center circle
