+++ other_imports/sigma/renderers/canvas/sigma.canvas.nodes.def.js
@@ -53,7 +53,7 @@ sigma.canvas.nodes.def = function (node, context, settings) {
     highlightNodeIfNecessary(context, node, size, x, y)
     context.lineWidth = lineWidth
 };
-function getDimensions(context, node, settings) {
+export function getDimensions(context, node, settings) {
     var prefix = settings('prefix') || '';
     const obj = {
         size: node[prefix + 'size'],
@@ -74,7 +74,7 @@ function drawPieSlice(ctx, centerX, centerY, radius, startAngle, endAngle, color
     ctx.fill();
 }
 
-function getColorFromNode(node) {
+export function getColorFromNode(node) {
     let color
   if(node && node.contentUserData && node.contentUserData.proficiency) {
       color = ProficiencyUtils.getColor(node.contentUserData.proficiency)
@@ -86,28 +86,10 @@ function getColorFromNode(node) {
 function getRectangleDimensions(node, x, y) {
     // return
 }
-function drawNodeRectangleNotFilled(context, node, size, x, y){
-    const halfWidth = size * 10
-    const height = calculateCardHeight(node, size)
-    const halfHeight = height / 2
-    const color = getColorFromNode(node)
-    context.lineWidth="4";
-    context.strokeStyle = color;
-    context.beginPath();
-    context.rect(
-        x - halfWidth,
-        y + halfHeight - 45,
-        halfWidth * 2,
-        height
-    );
-    context.stroke();
-}
-function drawNodeRectangleFilled(context, node, size, x, y) {
+export function drawNodeRectangleCore(context, node, size, x, y) {
     const halfWidth = size * 10;
     const height = calculateCardHeight(node, size)
     const halfHeight = height / 2
-    const color = getColorFromNode(node)
-    context.fillStyle = color;
     context.beginPath();
     context.rect(
         x - halfWidth,
@@ -115,6 +97,12 @@ function drawNodeRectangleFilled(context, node, size, x, y) {
         halfWidth * 2,
         height
     );
+
+}
+function drawNodeRectangleFilled(context, node, size, x, y) {
+    const color = getColorFromNode(node)
+    context.fillStyle = color;
+    drawNodeRectangleCore(context, node, size, x, y)
     // from https://github.com/jacomyal/sigma.js/wiki/Renderers
     context.closePath();
     context.fill();
@@ -138,7 +126,7 @@ function placeTextOnRectangle(context, node, x, y) {
     const text = "Hello this is a note"
 }
 
-function calculateCardHeight(node, size){
+export function calculateCardHeight(node, size){
     return size * 5;
     // return 50;
     //TODO: based off size of text
